+++ lib/crep/model/crash_sources/hockeyapp_crash_source.rb
@@ -15,7 +15,6 @@ module Crep
 
       @client = HockeyApp.build_client
       @hockeyapp_app = hockeyapp(bundle_identifier, @client)
-      CrepLogger.error("No app with the give bundle identifier (#{bundle_identifier}) was found.") unless @hockeyapp_app
       @app = App.new(@hockeyapp_app.title, bundle_identifier)
     end
 
@@ -53,7 +52,10 @@ module Crep
     def version(version:, build:)
       filtered_versions = filtered_versions_by_version_and_build(@hockeyapp_app.versions, version, build)
 
-      raise "No version was found for #{version})#{build})".red unless filtered_versions.count > 0
+      if filtered_versions.count <= 0
+        CrepLogger.error("No version was found for #{version})#{build})")
+        raise
+      end
 
       report_version = filtered_versions.first
 
@@ -70,6 +72,10 @@ module Crep
       all_apps = client.get_apps
       apps = all_apps.select do |app|
         app.bundle_identifier == bundle_identifier
+      end
+      if !apps.first
+        CrepLogger.error("No app with the give bundle identifier (#{bundle_identifier}) was found.") unless apps.first
+        raise
       end
       apps.first
     end